requirements.txt
mysql-python
django_nose